refactor(Profile): Clean up
